Create Snake,Water,Gun
Game of Snake, Water, Gun. I assume most of you have played this game in your childhood. And now We have grown up this far, that I have coded this game. Sounds thrilling right? The game tells you the instructions on how to play? and tell me if you win or not.